uploading and cropping functionality
