+++ components/add-entity-modal.tsx
@@ -8,6 +8,7 @@ import {
   DialogContent,
   DialogHeader,
   DialogTitle,
+  DialogFooter,
 } from "@/components/ui/dialog";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
@@ -20,13 +21,25 @@ import {
   SelectTrigger,
   SelectValue,
 } from "@/components/ui/select";
-import type { EntityType } from "@/app/admin/organisations/page";
+import ImageField from "./image-field"
+import type {
+  EntityType,
+  Organization,
+  Team,
+  Circle,
+  Person,
+} from "@/app/admin/organisations/page";
 
 interface AddEntityModalProps {
   isOpen: boolean;
   onClose: () => void;
   onAdd: (type: EntityType, data: any) => void;
-  data: any;
+  data: {
+    organizations: Organization[];
+    teams: Team[];
+    circles: Circle[];
+    people: Person[];
+  };
 }
 
 export default function AddEntityModal({
@@ -37,23 +50,24 @@ export default function AddEntityModal({
 }: AddEntityModalProps) {
   const [entityType, setEntityType] = useState<EntityType>("organization");
   const [formData, setFormData] = useState<any>({});
-  const [imagePreview, setImagePreview] = useState<string | null>(null);
 
   const handleSubmit = (e: React.FormEvent) => {
     e.preventDefault();
     onAdd(entityType, formData);
     setFormData({});
-    setImagePreview(null);
     onClose();
   };
 
-  const handleImageUrlChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const url = e.target.value;
-    setFormData({ ...formData, imageUrl: url });
-    setImagePreview(url || null);
+  const handleInputChange = (field: string, value: string) => {
+    setFormData((prev: any) => ({ ...prev, [field]: value }));
   };
 
-  const renderForm = () => {
+  const handleTypeChange = (type: EntityType) => {
+    setEntityType(type);
+    setFormData({});
+  };
+
+  const renderFormFields = () => {
     switch (entityType) {
       case "organization":
         return (
@@ -63,10 +77,8 @@ export default function AddEntityModal({
               <Input
                 id="name"
                 value={formData.name || ""}
-                onChange={(e) =>
-                  setFormData({ ...formData, name: e.target.value })
-                }
-                placeholder="e.g., DU1 - Digital Innovation"
+                onChange={(e) => handleInputChange("name", e.target.value)}
+                required
               />
             </div>
             <div className="space-y-2">
@@ -75,9 +87,8 @@ export default function AddEntityModal({
                 id="description"
                 value={formData.description || ""}
                 onChange={(e) =>
-                  setFormData({ ...formData, description: e.target.value })
+                  handleInputChange("description", e.target.value)
                 }
-                placeholder="Brief description of the organization"
               />
             </div>
             <div className="space-y-2">
@@ -85,21 +96,18 @@ export default function AddEntityModal({
               <Input
                 id="location"
                 value={formData.location || ""}
-                onChange={(e) =>
-                  setFormData({ ...formData, location: e.target.value })
-                }
-                placeholder="e.g., San Francisco, CA"
+                onChange={(e) => handleInputChange("location", e.target.value)}
               />
             </div>
             <div className="space-y-2">
               <Label htmlFor="established">Established</Label>
               <Input
                 id="established"
+                type="date"
                 value={formData.established || ""}
                 onChange={(e) =>
-                  setFormData({ ...formData, established: e.target.value })
+                  handleInputChange("established", e.target.value)
                 }
-                placeholder="e.g., 2020"
               />
             </div>
             <div className="space-y-2">
@@ -108,41 +116,17 @@ export default function AddEntityModal({
                 id="employees"
                 type="number"
                 value={formData.employees || ""}
-                onChange={(e) =>
-                  setFormData({
-                    ...formData,
-                    employees: Number.parseInt(e.target.value) || 0,
-                  })
-                }
-                placeholder="e.g., 150"
+                onChange={(e) => handleInputChange("employees", e.target.value)}
               />
             </div>
-            <div className="space-y-2">
-              <Label htmlFor="imageUrl">Profile Picture URL (Optional)</Label>
-              <div className="flex gap-2 items-center">
-                <Input
-                  id="imageUrl"
-                  value={formData.imageUrl || ""}
-                  onChange={handleImageUrlChange}
-                  placeholder="Leave blank for no image"
-                  className="flex-1"
-                />
-                {imagePreview && (
-                  <div className="h-10 w-10 rounded-full overflow-hidden border border-gray-200 flex-shrink-0">
-                    <img
-                      src={imagePreview || "/placeholder.svg"}
-                      alt="Preview"
-                      className="h-full w-full object-cover"
-                      onError={() => setImagePreview(null)}
-                    />
-                  </div>
-                )}
-              </div>
-              <p className="text-xs text-gray-500">
-                Enter a URL to a profile picture or leave blank if no image is
-                needed.
-              </p>
-            </div>
+            <ImageField
+              label="Organization Image"
+              value={formData.imageUrl || ""}
+              onChange={(value) =>
+                setFormData({ ...formData, imageUrl: value })
+              }
+              aspectRatio={16 / 9}
+            />
           </>
         );
 
@@ -154,14 +138,14 @@ export default function AddEntityModal({
               <Select
                 value={formData.organizationId || ""}
                 onValueChange={(value) =>
-                  setFormData({ ...formData, organizationId: value })
+                  handleInputChange("organizationId", value)
                 }
               >
                 <SelectTrigger>
                   <SelectValue placeholder="Select organization" />
                 </SelectTrigger>
                 <SelectContent>
-                  {data.organizations.map((org: any) => (
+                  {data.organizations.map((org) => (
                     <SelectItem key={org.id} value={org.id}>
                       {org.name}
                     </SelectItem>
@@ -174,10 +158,8 @@ export default function AddEntityModal({
               <Input
                 id="name"
                 value={formData.name || ""}
-                onChange={(e) =>
-                  setFormData({ ...formData, name: e.target.value })
-                }
-                placeholder="e.g., Team Alpha"
+                onChange={(e) => handleInputChange("name", e.target.value)}
+                required
               />
             </div>
             <div className="space-y-2">
@@ -186,9 +168,8 @@ export default function AddEntityModal({
                 id="description"
                 value={formData.description || ""}
                 onChange={(e) =>
-                  setFormData({ ...formData, description: e.target.value })
+                  handleInputChange("description", e.target.value)
                 }
-                placeholder="What does this team do?"
               />
             </div>
             <div className="space-y-2">
@@ -196,10 +177,7 @@ export default function AddEntityModal({
               <Input
                 id="lead"
                 value={formData.lead || ""}
-                onChange={(e) =>
-                  setFormData({ ...formData, lead: e.target.value })
-                }
-                placeholder="e.g., Sarah Johnson"
+                onChange={(e) => handleInputChange("lead", e.target.value)}
               />
             </div>
             <div className="space-y-2">
@@ -207,38 +185,17 @@ export default function AddEntityModal({
               <Input
                 id="focus"
                 value={formData.focus || ""}
-                onChange={(e) =>
-                  setFormData({ ...formData, focus: e.target.value })
-                }
-                placeholder="e.g., User Interface"
+                onChange={(e) => handleInputChange("focus", e.target.value)}
               />
             </div>
-            <div className="space-y-2">
-              <Label htmlFor="imageUrl">Profile Picture URL (Optional)</Label>
-              <div className="flex gap-2 items-center">
-                <Input
-                  id="imageUrl"
-                  value={formData.imageUrl || ""}
-                  onChange={handleImageUrlChange}
-                  placeholder="Leave blank for no image"
-                  className="flex-1"
-                />
-                {imagePreview && (
-                  <div className="h-10 w-10 rounded-full overflow-hidden border border-gray-200 flex-shrink-0">
-                    <img
-                      src={imagePreview || "/placeholder.svg"}
-                      alt="Preview"
-                      className="h-full w-full object-cover"
-                      onError={() => setImagePreview(null)}
-                    />
-                  </div>
-                )}
-              </div>
-              <p className="text-xs text-gray-500">
-                Enter a URL to a profile picture or leave blank if no image is
-                needed.
-              </p>
-            </div>
+            <ImageField
+              label="Team Image"
+              value={formData.imageUrl || ""}
+              onChange={(value) =>
+                setFormData({ ...formData, imageUrl: value })
+              }
+              aspectRatio={1}
+            />
           </>
         );
 
@@ -249,15 +206,13 @@ export default function AddEntityModal({
               <Label htmlFor="teamId">Team</Label>
               <Select
                 value={formData.teamId || ""}
-                onValueChange={(value) =>
-                  setFormData({ ...formData, teamId: value })
-                }
+                onValueChange={(value) => handleInputChange("teamId", value)}
               >
                 <SelectTrigger>
                   <SelectValue placeholder="Select team" />
                 </SelectTrigger>
                 <SelectContent>
-                  {data.teams.map((team: any) => (
+                  {data.teams.map((team) => (
                     <SelectItem key={team.id} value={team.id}>
                       {team.name}
                     </SelectItem>
@@ -270,10 +225,8 @@ export default function AddEntityModal({
               <Input
                 id="name"
                 value={formData.name || ""}
-                onChange={(e) =>
-                  setFormData({ ...formData, name: e.target.value })
-                }
-                placeholder="e.g., UI/UX Circle"
+                onChange={(e) => handleInputChange("name", e.target.value)}
+                required
               />
             </div>
             <div className="space-y-2">
@@ -281,10 +234,7 @@ export default function AddEntityModal({
               <Textarea
                 id="purpose"
                 value={formData.purpose || ""}
-                onChange={(e) =>
-                  setFormData({ ...formData, purpose: e.target.value })
-                }
-                placeholder="What is the purpose of this circle?"
+                onChange={(e) => handleInputChange("purpose", e.target.value)}
               />
             </div>
             <div className="space-y-2">
@@ -295,43 +245,21 @@ export default function AddEntityModal({
                 id="responsibilities"
                 value={formData.responsibilities?.join(", ") || ""}
                 onChange={(e) =>
-                  setFormData({
-                    ...formData,
-                    responsibilities: e.target.value
-                      .split(",")
-                      .map((r) => r.trim())
-                      .filter((r) => r),
-                  })
+                  handleInputChange(
+                    "responsibilities",
+                    e.target.value.split(", ").filter(Boolean)
+                  )
                 }
-                placeholder="e.g., User Research, Interface Design, Prototyping"
               />
             </div>
-            <div className="space-y-2">
-              <Label htmlFor="imageUrl">Profile Picture URL (Optional)</Label>
-              <div className="flex gap-2 items-center">
-                <Input
-                  id="imageUrl"
-                  value={formData.imageUrl || ""}
-                  onChange={handleImageUrlChange}
-                  placeholder="Leave blank for no image"
-                  className="flex-1"
-                />
-                {imagePreview && (
-                  <div className="h-10 w-10 rounded-full overflow-hidden border border-gray-200 flex-shrink-0">
-                    <img
-                      src={imagePreview || "/placeholder.svg"}
-                      alt="Preview"
-                      className="h-full w-full object-cover"
-                      onError={() => setImagePreview(null)}
-                    />
-                  </div>
-                )}
-              </div>
-              <p className="text-xs text-gray-500">
-                Enter a URL to a profile picture or leave blank if no image is
-                needed.
-              </p>
-            </div>
+            <ImageField
+              label="Circle Image"
+              value={formData.imageUrl || ""}
+              onChange={(value) =>
+                setFormData({ ...formData, imageUrl: value })
+              }
+              aspectRatio={1}
+            />
           </>
         );
 
@@ -339,58 +267,16 @@ export default function AddEntityModal({
         return (
           <>
             <div className="space-y-2">
-              <Label htmlFor="name">Name</Label>
-              <Input
-                id="name"
-                value={formData.name || ""}
-                onChange={(e) =>
-                  setFormData({ ...formData, name: e.target.value })
-                }
-              />
-            </div>
-            <div className="space-y-2">
-              <Label htmlFor="role">Role</Label>
-              <Input
-                id="role"
-                value={formData.role || ""}
-                onChange={(e) =>
-                  setFormData({ ...formData, role: e.target.value })
-                }
-              />
-            </div>
-            <div className="space-y-2">
-              <Label htmlFor="email">Email</Label>
-              <Input
-                id="email"
-                value={formData.email || ""}
-                onChange={(e) =>
-                  setFormData({ ...formData, email: e.target.value })
-                }
-              />
-            </div>
-            <div className="space-y-2">
-              <Label htmlFor="department">Department</Label>
-              <Input
-                id="department"
-                value={formData.department || ""}
-                onChange={(e) =>
-                  setFormData({ ...formData, department: e.target.value })
-                }
-              />
-            </div>
-            <div className="space-y-2">
               <Label htmlFor="circleId">Circle</Label>
               <Select
                 value={formData.circleId || ""}
-                onValueChange={(value) =>
-                  setFormData({ ...formData, circleId: value })
-                }
+                onValueChange={(value) => handleInputChange("circleId", value)}
               >
                 <SelectTrigger>
                   <SelectValue placeholder="Select circle" />
                 </SelectTrigger>
                 <SelectContent>
-                  {data.circles.map((circle: any) => (
+                  {data.circles.map((circle) => (
                     <SelectItem key={circle.id} value={circle.id}>
                       {circle.name}
                     </SelectItem>
@@ -399,31 +285,49 @@ export default function AddEntityModal({
               </Select>
             </div>
             <div className="space-y-2">
-              <Label htmlFor="imageUrl">Profile Picture URL (Optional)</Label>
-              <div className="flex gap-2 items-center">
-                <Input
-                  id="imageUrl"
-                  value={formData.imageUrl || ""}
-                  onChange={handleImageUrlChange}
-                  placeholder="Leave blank for no image"
-                  className="flex-1"
-                />
-                {imagePreview && (
-                  <div className="h-10 w-10 rounded-full overflow-hidden border border-gray-200 flex-shrink-0">
-                    <img
-                      src={imagePreview || "/placeholder.svg"}
-                      alt="Preview"
-                      className="h-full w-full object-cover"
-                      onError={() => setImagePreview(null)}
-                    />
-                  </div>
-                )}
-              </div>
-              <p className="text-xs text-gray-500">
-                Enter a URL to a profile picture or leave blank if no image is
-                needed.
-              </p>
+              <Label htmlFor="name">Full Name</Label>
+              <Input
+                id="name"
+                value={formData.name || ""}
+                onChange={(e) => handleInputChange("name", e.target.value)}
+                required
+              />
             </div>
+            <div className="space-y-2">
+              <Label htmlFor="role">Role</Label>
+              <Input
+                id="role"
+                value={formData.role || ""}
+                onChange={(e) => handleInputChange("role", e.target.value)}
+              />
+            </div>
+            <div className="space-y-2">
+              <Label htmlFor="email">Email</Label>
+              <Input
+                id="email"
+                type="email"
+                value={formData.email || ""}
+                onChange={(e) => handleInputChange("email", e.target.value)}
+              />
+            </div>
+            <div className="space-y-2">
+              <Label htmlFor="department">Department</Label>
+              <Input
+                id="department"
+                value={formData.department || ""}
+                onChange={(e) =>
+                  handleInputChange("department", e.target.value)
+                }
+              />
+            </div>
+            <ImageField
+              label="Profile Picture"
+              value={formData.imageUrl || ""}
+              onChange={(value) =>
+                setFormData({ ...formData, imageUrl: value })
+              }
+              aspectRatio={1}
+            />
           </>
         );
     }
@@ -431,7 +335,7 @@ export default function AddEntityModal({
 
   return (
     <Dialog open={isOpen} onOpenChange={onClose}>
-      <DialogContent className="max-w-md max-h-[80vh] overflow-y-auto">
+      <DialogContent className="max-w-md max-h-[90vh] overflow-y-auto">
         <DialogHeader>
           <DialogTitle>Add New Entity</DialogTitle>
         </DialogHeader>
@@ -439,14 +343,7 @@ export default function AddEntityModal({
         <form onSubmit={handleSubmit} className="space-y-4">
           <div className="space-y-2">
             <Label htmlFor="entityType">Entity Type</Label>
-            <Select
-              value={entityType}
-              onValueChange={(value: EntityType) => {
-                setEntityType(value);
-                setFormData({});
-                setImagePreview(null);
-              }}
-            >
+            <Select value={entityType} onValueChange={handleTypeChange}>
               <SelectTrigger>
                 <SelectValue />
               </SelectTrigger>
@@ -459,14 +356,14 @@ export default function AddEntityModal({
             </Select>
           </div>
 
-          {renderForm()}
+          {renderFormFields()}
 
-          <div className="flex justify-end gap-2 pt-4">
+          <DialogFooter>
             <Button type="button" variant="outline" onClick={onClose}>
               Cancel
             </Button>
             <Button type="submit">Add {entityType}</Button>
-          </div>
+          </DialogFooter>
         </form>
       </DialogContent>
     </Dialog>
