+++ package.json
@@ -21,6 +21,7 @@
     "next": "15.2.4",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-image-crop": "^11.0.10",
     "tailwind-merge": "^3.3.0"
   },
   "devDependencies": {
